// src/types.ts
export interface ButtonViewReturnComponentProps {
  /** Method triggered when action is performed */
  action?: (value?: any) => void
  /** Whether it is in the active state */
  isActive?: () => boolean
  /** Button icon */
  icon?: any
  /** Text displayed on hover */
  tooltip?: string
  [x: string]: any
}